Split SQLite thread store SQL helpers
